can now generate a new population

+++ player.h
@@ -3,7 +3,7 @@
 
 
 struct player {
-	int offer;
+	double offer;
 	double lbound;
 	double ubound;
 };
